Branch kissel

+++ src/main/java/com/eventoapp/Interfaces/EventoRepository.java
@@ -1,9 +1,11 @@
 package com.eventoapp.Interfaces;
 
 import com.eventoapp.Entity.Evento;
-
 import org.springframework.data.repository.CrudRepository;
 
 public interface EventoRepository extends CrudRepository<Evento, String> {
-    Evento findByCodigo(long codigo);
+    Evento findByCodigo(int codigo);
+
+    // TO-DO versão 2.0
+    // Evento findAll();
 }
